# ContentService/docs/openapi.yaml
openapi: 3.0.3
info:
  title: PocketLOL Content Service API
  version: 1.0.0
  description: >-
    HTTP contract for viewer- and admin-facing catalog endpoints exposed through API Gateway.
servers:
  - url: https://api.pocketlol/v1
    description: Production gateway
  - url: https://staging.api.pocketlol/v1
    description: Staging gateway
paths:
...
  /catalog/series/{slug}:
    get:
      operationId: getSeriesDetail
      summary: Fetch full detail for a series, including localized metadata and availability.
      tags:
        - Viewer Catalog
      parameters:
        - name: slug
          in: path
          required: true
          description: Series slug used in viewer experiences.
          schema:
            type: string
      responses:
        '200':
          description: Series detail and related episode collections.
          headers:
            Cache-Control:
              schema:
                type: string
            X-Cache:
              schema:
                type: string
                enum: [hit, miss]
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeriesDetailResponse'
        '404':
          description: Series not found.
...
  /admin/catalog/series:
    post:
      operationId: createSeries
      summary: Create a series with hierarchical metadata and scheduling controls.
      tags:
        - Admin Catalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSeriesInput'
      responses:
        '201':
          description: Series created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Series'
        '409':
          description: Duplicate slug detected.
...
components:
  schemas:
...
    CatalogFeedItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        title:
          type: string
        synopsis:
          type: string
          nullable: true
        heroImageUrl:
          type: string
          format: uri
          nullable: true
        defaultThumbnailUrl:
          type: string
          format: uri
          nullable: true
        durationSeconds:
          type: integer
          minimum: 0
        publishedAt:
          type: string
          format: date-time
        availability:
          type: object
          properties:
            start:
              type: string
              format: date-time
              nullable: true
            end:
              type: string
              format: date-time
              nullable: true
          required: [start, end]
        season:
          $ref: '#/components/schemas/CatalogSeasonSummary'
        series:
          $ref: '#/components/schemas/CatalogSeriesSummary'
        playback:
          $ref: '#/components/schemas/CatalogPlayback'
        localization:
          $ref: '#/components/schemas/CatalogLocalization'
        personalization:
          $ref: '#/components/schemas/CatalogPersonalization'
        ratings:
          type: object
          properties:
            average:
              type: number
              nullable: true
          required: [average]
      required:
        - id
        - slug
        - title
        - durationSeconds
        - publishedAt
        - availability
        - series
        - playback
        - localization
        - personalization
        - ratings
    CatalogSeasonSummary:
      type: object
      nullable: true
      properties:
        id:
          type: string
          format: uuid
        sequenceNumber:
          type: integer
        title:
          type: string
          nullable: true
      required: [id, sequenceNumber, title]
    CatalogSeriesSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        title:
          type: string
        synopsis:
          type: string
          nullable: true
        heroImageUrl:
          type: string
          format: uri
          nullable: true
        bannerImageUrl:
          type: string
          format: uri
          nullable: true
        category:
          $ref: '#/components/schemas/CategorySummary'
      required: [id, slug, title]
    CatalogPlayback:
      type: object
      properties:
        status:
          type: string
          enum: [PENDING, PROCESSING, READY, FAILED]
        manifestUrl:
          type: string
          format: uri
          nullable: true
        defaultThumbnailUrl:
          type: string
          format: uri
          nullable: true
        variants:
          type: array
          items:
            $ref: '#/components/schemas/PlaybackVariant'
      required: [status, variants]
    PlaybackVariant:
      type: object
      properties:
        label:
          type: string
        width:
          type: integer
          nullable: true
        height:
          type: integer
          nullable: true
        bitrateKbps:
          type: integer
          nullable: true
        codec:
          type: string
          nullable: true
        frameRate:
          type: number
          nullable: true
      required: [label]
    CatalogLocalization:
      type: object
      properties:
        captions:
          type: array
          items:
            type: object
            properties:
              language:
                type: string
              label:
                type: string
                nullable: true
              url:
                type: string
                format: uri
                nullable: true
            required: [language]
        availableLanguages:
          type: array
          items:
            type: string
      required: [captions, availableLanguages]
    CatalogPersonalization:
      type: object
      properties:
        reason:
          type: string
          enum: [trending, recent, viewer_following]
        score:
          type: number
          nullable: true
      required: [reason]
    SeriesDetailResponse:
      type: object
      properties:
        series:
          $ref: '#/components/schemas/Series'
        seasons:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              sequenceNumber:
                type: integer
              title:
                type: string
              synopsis:
                type: string
                nullable: true
              releaseDate:
                type: string
                format: date-time
                nullable: true
              episodes:
                type: array
                items:
                  $ref: '#/components/schemas/CatalogFeedItem'
            required: [id, sequenceNumber, title, episodes]
        standaloneEpisodes:
          type: array
          items:
            $ref: '#/components/schemas/CatalogFeedItem'
      required: [series, seasons, standaloneEpisodes]
...
    CategorySummary:
      type: object
      nullable: true
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
      required: [id, slug, name]
...
    Series:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        title:
          type: string
        synopsis:
          type: string
          nullable: true
        heroImageUrl:
          type: string
          format: uri
          nullable: true
        bannerImageUrl:
          type: string
          format: uri
          nullable: true
        tags:
          type: array
          items:
            type: string
        status:
          $ref: '#/components/schemas/PublicationStatus'
        visibility:
          $ref: '#/components/schemas/Visibility'
        releaseDate:
          type: string
          format: date-time
          nullable: true
        categoryId:
          type: string
          format: uuid
          nullable: true
      required:
        - id
        - slug
        - title
        - tags
        - status
        - visibility
    AdminSeriesInput:
      type: object
      required: [slug, title]
      properties:
        slug:
          type: string
        title:
          type: string
        synopsis:
          type: string
          nullable: true
        heroImageUrl:
          type: string
          format: uri
          nullable: true
        bannerImageUrl:
          type: string
          format: uri
          nullable: true
        tags:
          type: array
          items:
            type: string
        status:
          $ref: '#/components/schemas/PublicationStatus'
        visibility:
          $ref: '#/components/schemas/Visibility'
        releaseDate:
          type: string
          format: date-time
          nullable: true
        ownerId:
          type: string
          format: uuid
          nullable: true
        categoryId:
          type: string
          format: uuid
          nullable: true
...
    PublicationStatus:
      type: string
      enum: [DRAFT, REVIEW, PUBLISHED, ARCHIVED]
    Visibility:
      type: string
      enum: [PUBLIC, UNLISTED, PRIVATE]